Remove footer
The footer div is still there in case we want to fill it in later. But I
removed the mention of Sphinx since the page has been hacked up since
then, and Sphinx cannot be used to regenerate the page since the
original source doc is not (AFAICT) in the Git repository.