Bio-Formats does not depend on SCIFIO anymore
The definition of "SCIFIO" evolved over the years. The SCIFIO project is
now part of the SciJava software stack, and a key component of ImageJ2.

The BSD-licensed core of Bio-Formats is no longer called SCIFIO,
but rather just the "formats-api" and "formats-bsd" components.

SCIFIO and Bio-Formats are integrated via the scifio-bf-compat
component (https://github.com/scifio/scifio-bf-compat).