Remove the docs folder
It was (I believe) originally used to generate the static site.
However, the static site itself is also committed here.

We should either:

A) Stick to editing a static site (what we are doing now); or:
B) Switch to Jekyll and let GitHub regenerate the site for us.

For now, we'll keep doing (A). Soon, I hope to redesign the front page,
and when that happens, we can seriously consider going to Jekyll.